requirements.txt
ansible==12.1.0
ansible-core==2.19.3
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ncclient==0.7.0
netmiko==4.6.0
ntc_templates==8.1.0
packaging==25.0
paramiko==4.0.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.0
pyserial==3.5
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
resolvelib==1.2.1
rich==14.2.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scp==0.15.0
textfsm==2.1.0
urllib3==2.5.0
xmltodict==1.0.2
